Add feature for stopping campaign

+++ src/main/java/com/pet/mailSender/dao/imp/AbstractDao.java
@@ -1,12 +1,18 @@
 package com.pet.mailSender.dao.imp;
 
 import com.pet.mailSender.dao.Dao;
+import org.hibernate.Transaction;
+import org.springframework.stereotype.Repository;
+import sun.rmi.transport.Transport;
+
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.List;
 
+@Repository
 public abstract class AbstractDao <T extends Serializable> implements Dao<T> {
 
     private EntityManagerFactory entityManagerFactory = Persistence.createEntityManagerFactory("com.pet.mailSender.model");
@@ -20,40 +26,79 @@ public abstract class AbstractDao <T extends Serializable> implements Dao<T> {
 
     @Override
     public void add(T entity) {
-        entityManager.getTransaction().begin();
-        entityManager.persist(entity);
-        entityManager.getTransaction().commit();
+        if(!entityManager.getTransaction().isActive()){
+            try{
+                entityManager.getTransaction().begin();
+                entityManager.persist(entity);
+                entityManager.getTransaction().commit();
+            }catch (Exception e){
+               // entityManager.getTransaction().rollback();
+            }
+
+        }
+
     }
 
     @Override
     public List<T> getAll() {
-        entityManager.getTransaction().begin();
-        List<T> result = entityManager.createQuery("FROM " + clazz.getName(), clazz).getResultList();
-        entityManager.getTransaction().commit();
+        List<T> result = new ArrayList<>();
+        if(!entityManager.getTransaction().isActive()) {
+            try{
+                entityManager.getTransaction().begin();
+                result = entityManager.createQuery("FROM " + clazz.getName(), clazz).getResultList();
+                entityManager.getTransaction().commit();
+            }catch (Exception e){
+               // entityManager.getTransaction().rollback();
+            }
+
+        }
         return result;
     }
 
     @Override
     public T getById(int id) {
-        entityManager.getTransaction().begin();
-        T object = entityManager.find(clazz, id);
-        entityManager.getTransaction().commit();
+        T object = null;
+        if(!entityManager.getTransaction().isActive()){
+            try{
+                entityManager.getTransaction().begin();
+                object = entityManager.find(clazz, id);
+                entityManager.getTransaction().commit();
+            }catch (Exception e){
+               // entityManager.getTransaction().rollback();
+            }
+
+        }
         return object;
     }
 
     @Override
     public T update(T entity) {
-        entityManager.getTransaction().begin();
-        T object = entityManager.merge(entity);
-        entityManager.getTransaction().commit();
+        T object = null;
+        if(!entityManager.getTransaction().isActive()) {
+            try{
+                entityManager.getTransaction().begin();
+                object = entityManager.merge(entity);
+                entityManager.getTransaction().commit();
+            }catch (Exception e){
+               // entityManager.getTransaction().rollback();
+            }
+
+        }
         return object;
     }
 
     @Override
     public void delete(T entity) {
-        entityManager.getTransaction().begin();
-        entityManager.remove(entity);
-        entityManager.getTransaction().commit();
+        if(!entityManager.getTransaction().isActive()) {
+            try{
+                entityManager.getTransaction().begin();
+                entityManager.remove(entity);
+                entityManager.getTransaction().commit();
+            }catch (Exception e){
+               // entityManager.getTransaction().rollback();
+            }
+
+        }
     }
 
 }
